+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * UNRELEASED
   * Added support for nth-day-of-the-week holidays
+  * Now supports observed holidays
 
 * 0.5.0 - Jan 7, 2022
   * Upgraded to tick 0.4.32